Fix performance problem and false positives.
The performance problem was due to a Cartesian product, the false positives due to overly stringent requirements on the various checks happening on exactly the same SSA-with-fields variable. I have relaxed the latter (in a technically unsound way) to gloss over differences in the SSA graph. This can, but is unlikely to, cause false negatives.