Rename ql/test/query-tests/Security/CWE-601/BadRedirectCheck/main.go to ql/test/query-tests/Security/CWE-601/BadRedirectCheck/ramos/main.go
Redirect URLs should be checked to ensure that user input cannot cause a site to redirect to arbitrary domains. This is often done with a check that the redirect URL begins with a slash, which most of the time is an absolute redirect on the same host. However, browsers interpret URLs beginning with // or /\ as absolute URLs. For example, a redirect to //lgtm.com will redirect to https://lgtm.com. Thus, redirect checks must also check the second character of redirect URLs.